exa: more work on composite
Getting closer.. still need to figure out how to deal w/ porter-duff op
encoding, which seems to also be a function of src and dst (but not mask?)
formats.

Also, need to figure out how to deal with maskX/maskY.. libC2D2 does not
give a way to specify different src and mask rect, so not quite sure if
this is a hw limit.. it might be worth experimenting w/ cmdstream to see
if we can guess how to handle this.  Otherwise, we might be able to fake
it via ptr arithmatic on the gpuaddr.