accel: initial support for submitting multiple blits
Performance really sucks if you don't batch up multiple blits to
submit at the same time.  The next step, with the help of some
changes on the kernel side, would be make it so that we don't
have to wait for the blits to complete before building up the
cmdstream for the next blits, although this is a bit awkward
with the current way that context-restore works (see NOTES in
freedreno tree).. this can be solved by re-working the kernel
side to allow issueibcmds to take the ctxt restore packet and
the blit cmds as two separate buffers.

But the first issue that needs to be solved is the issue of
multiple blit-complete interrupts (or rather single interrupt
with count>1.  This results in "complete" timestamp incrementing
past the "pending" timestamp, so the driver things blits that
are still queued or in progress are complete.  Which causes all
sorts of issues, like thinking we can free pixmaps which the
GPU is still blitting to/from.