accel: reduce cmdstream buffer size
The size field in VGV3_NEXTCMD is only 12 bits, so we can submit
cmdstream packets with payload larger than this.  Reduce the size
of the cmdstream buffer to compensate.

Also, the smaller size turns up some bugs in handling of ringbuffer
switch (we'd switch in BEGIN_RING() -> FIRE_RING(), but the exa fxns
grab the ringbuffer ptr before that), so fix that as well.